Reduce history entry sizes as the resolution is only 14-bit wide (#34)
ELO   | 0.85 +- 2.88 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.98 (-2.94, 2.94) [-4.00, 1.00]
GAMES | N: 24512 W: 5410 L: 5350 D: 13752
http://chess.grantnet.us/test/19532

Bench: 10,366,635

+++ src/include/history.h
@@ -31,16 +31,16 @@ enum
     HistoryResolution = HistoryMaxScore * HistoryScale
 };
 
-typedef int32_t butterfly_history_t[COLOR_NB][SQUARE_NB * SQUARE_NB];
-typedef int32_t piece_history_t[PIECE_NB][SQUARE_NB];
+typedef int16_t butterfly_history_t[COLOR_NB][SQUARE_NB * SQUARE_NB];
+typedef int16_t piece_history_t[PIECE_NB][SQUARE_NB];
 typedef piece_history_t continuation_history_t[PIECE_NB][SQUARE_NB];
 typedef move_t countermove_history_t[PIECE_NB][SQUARE_NB];
 
 INLINED void add_bf_history(butterfly_history_t hist, piece_t piece, move_t move, int32_t bonus)
 {
-    int32_t *entry = &hist[piece_color(piece)][square_mask(move)];
+    int16_t *entry = &hist[piece_color(piece)][square_mask(move)];
 
-    *entry += bonus - *entry * abs(bonus) / HistoryResolution;
+    *entry += bonus - (int32_t)*entry * abs(bonus) / HistoryResolution;
 }
 
 INLINED score_t get_bf_history_score(const butterfly_history_t hist, piece_t piece, move_t move)
@@ -50,9 +50,9 @@ INLINED score_t get_bf_history_score(const butterfly_history_t hist, piece_t pie
 
 INLINED void add_pc_history(piece_history_t hist, piece_t pc, square_t to, int32_t bonus)
 {
-    int32_t *entry = &hist[pc][to];
+    int16_t *entry = &hist[pc][to];
 
-    *entry += bonus - *entry * abs(bonus) / HistoryResolution;
+    *entry += bonus - (int32_t)*entry * abs(bonus) / HistoryResolution;
 }
 
 INLINED score_t get_pc_history_score(const piece_history_t hist, piece_t pc, square_t to)
